notify-api/pyproject.toml
[tool.poetry]
name = "notify-api"
version = "2.0.1"
description = ""
authors = ["pwei <patrick.wei@gov.bc.ca>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "notify_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
flask-babel = "^3.1.0"
Flask-Cors = "^3.0.10"
flask-migrate = "^4.0.4"
psycopg2 = "^2.9.6"
launchdarkly-server-sdk = "^8.1.3"
python-dotenv = "^1.0.0"
Flask-Pydantic = "^0.11.0"
PyPDF2 = "^3.0.1"
attrs = "^23.1.0"
flask-jwt-oidc = "^0.3.0"
datedelta = "^1.4"
dpath = "^2.1.6"
gunicorn = "^20.1.0"
jsonschema = "^4.17.3"
pycountry = "^22.3.5"
pydantic = "^1.10.8"
reportlab = "^4.0.0"
requests = "^2.31.0"
strict-rfc3339 = "^0.7"
email-validator= "1.3.1"
phonenumbers = "8.13.6"
notifications-python-client = "7.0.0"
beautifulsoup4 = "4.11.2"
sqlalchemy-continuum = "^1.3.14"
pg8000 = "^1.30.3"

[tool.poetry.group.test.dependencies]
freezegun = "^1.2.2"
hypothesis = "^6.75.6"
lovely-pytest-docker = "^0.3.1"
pexpect = "^4.8.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
requests-mock = "^1.11.0"
faker = "^19.12.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pylint = "^3.0.3"
bandit = "^1.7.6"
flake8-pyproject = "^1.2.3"
isort = "^5.13.2"

[tool.bandit]
exclude_dirs = [".venv","tests"]
skips = ["B104"]

[tool.flake8]
ignore = ["F401","E402", "Q000", "E203"]
exclude = [
    ".venv",
    ".git",
    "*migrations*"
]
per-file-ignores = [
    "__init__.py:F401",
    "*.py:B902"
]
max-line-length = 120
docstring-min-length=10
count = true

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pylint.'MESSAGES CONTROL']
ignore = [ "migrations", "test"]
notes = ["FIXME","XXX","TODO"]
ignored-modules= ["flask_sqlalchemy", "sqlalchemy", "SQLAlchemy" , "alembic", "scoped_session"]
ignored-classes= "scoped_session"
disable = "C0209,C0301,W0511,W0613,W0703,W1514,R0801,R0902,R0903,R0911,R0401,R1705,R1718,W3101"
extension-pkg-whitelist = "pydantic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"